@@ -10,7 +10,7 @@
     <properties>
         <revision>1020.121.0</revision>
         <changelist></changelist>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
         <commons-lang3.version>3.11</commons-lang3.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <nexus.url>http://localhost:8080</nexus.url>
