@@ -38,7 +38,7 @@
         <changelist></changelist>
         <cumulocity-dependencies-tools.version>1020.40.0</cumulocity-dependencies-tools.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
         <nexus.url>http://localhost:8080</nexus.url>
         <nexus.basePath>/nexus/content/repositories</nexus.basePath>
         <aspectj.version>1.8.14</aspectj.version>
@@ -48,7 +48,7 @@
         <guava.version>32.0.1-jre</guava.version>
         <hamcrest.version>2.2</hamcrest.version>
         <httpclient.version>4.5.13</httpclient.version>
-        <jaxrs.version>2.1.6</jaxrs.version>
+        <jaxrs.version>3.1.0</jaxrs.version>
         <jetty.version>9.4.53.v20231009</jetty.version>
         <jsonassert.version>1.5.1</jsonassert.version>
         <junit.version>5.8.2</junit.version>
@@ -59,7 +59,7 @@
         <mockito.version>4.5.1</mockito.version>
         <slf4j.version>1.7.36</slf4j.version>
         <spring.version>${spring-framework.version}</spring.version><!-- deprecated, use spring-framework.version instead -->
-        <spring-framework.version>5.3.24</spring-framework.version>
+        <spring-framework.version>6.0.12</spring-framework.version>
         <svenson.version>1.5.8</svenson.version>
         <commons-lang3.version>3.11</commons-lang3.version>
         <commons-io.version>2.8.0</commons-io.version>
@@ -509,7 +509,7 @@
             <dependency>
                 <groupId>org.glassfish.jersey</groupId>
                 <artifactId>jersey-bom</artifactId>
-                <version>2.35</version>
+                <version>3.1.3</version>
                 <scope>import</scope>
                 <type>pom</type>
             </dependency>
