@@ -2,8 +2,8 @@ package com.cumulocity.sdk.client;
 
 import java.io.IOException;
 
-import javax.ws.rs.client.ClientRequestContext;
-import javax.ws.rs.client.ClientRequestFilter;
+import jakarta.ws.rs.client.ClientRequestContext;
+import jakarta.ws.rs.client.ClientRequestFilter;
 import java.nio.charset.StandardCharsets;
 import java.util.Base64;
 
