@@ -1,7 +1,7 @@
 package com.cumulocity.sdk.client.interceptor;
 
 
-import javax.ws.rs.client.Invocation;
+import jakarta.ws.rs.client.Invocation;
 
 /**
  * Interceptor will ce called before http method will be executed
